requirements.txt
transformers>=4.44.0
trl>=0.11.0
accelerate>=0.33.0
peft>=0.12.0
datasets>=2.20.0
bitsandbytes>=0.43.0
mlflow>=2.15.0
fastapi>=0.112.0
uvicorn>=0.30.0
pyyaml>=6.0.1
deepspeed>=0.15.0
